Move from es6 map to a object type map and don't use replicants for configs/clients
ES6 Maps don't play nicely with NodeCG Replicants, as they can't be serialized using JSON because they rely on iterators. Therefore we can't access them using the gui which was the reason to move to javascript object maps that can be transmitted to the gui using NodeCG Replicants.

As a consequence of this we can't use Replicants for service instance configs/clients anymore because NodeCG restricts having a object in more than one Replicant accessible. In this case you could access them through the service instance Replicant and the Replicant of each client/config. Therefore I also had to ditch those, store them normally and have callbacks that let the bundle manager know once a client of a service instance updates. As a nice site-improvement this also vastly simplifies persistence once we get there as there are no duplicate objects anymore (e.g. service and service through service instance).