Don't save services in replicants
Followup on #90

When saving services in Replicants other bundles have read and write access. I've assumed in #90 that thats fine based on my understanding that replicants always share data through serialization thus functions of the service objects aren't available to other bundles.
Turns out that this is not the case for server-side executed extensions, those get full access. Therefore a bundle could overwrite the createClient or validateConfig function to gain access to the passed config/credentials.
Because of that this PR also removes the services from Replicants and moves them into a array and gives out copies without the functions via nodecg messages.