Properly format obs client creation errors in createClient
The obs client throws a object when a error happens. When the framework catches the error it just tries to convert it to a string but when doing this with a object it just produces `[object Object]`.
This PR catches the exception and extracts the actual error message of the object to solve this, as it is also done in the `validateConfig` method.