Simplify importing of core
Adds a `index.ts` to core that re-exports everything important. That way services and bundles can import everything they need by just doing this:
```typescript
import { ... } from "nodecg-io-core";
```

Previously bundle/service developers would need to know in which file is which class and provide the full path.
Everything obviously still can be imported by using the full path but if anything is moved this doesn't need fixing of the paths and therefore is prefered.

This is a follow up on #131 which simplified importing of services.