Save spotify refresh token
Saves the spotify refresh token so that it can be re-used to create a access token at startup. That way we only need to open a tab to login once (the first start) instead of every time.
This can also be used as a workarround to get spotify working on a headless server.