Don't publish dashboard with core
When packing nodecg-io-core it would include the dashboard, which has a symlink to the core in the node-modules directory which resulted in a infinte loop when trying to pack the core.
The dashboard will now also be published to npm and in the planed yeoman generator, which will install releases of nodecg-io, will download it separately.