Warn on console.log usage and embrace nodecg logger
NodeCG bundles should use the integrated nodecg logger and while developing you easily forget it and use `console.log`/`console.error` which may be forgotten and left in the code.
This also introduces a inconsistency because sometimes the nodecg logger and sometimes the `console.log` is used.
To deal with this, this PR adds a eslint rule to treat `console.log` as a warning to ensure that the nodecg logger is used before code is merged. While developing and trying things out you should be still able to use `console.log` if you wish to, that's why only a warning and not an error.