Refactor ServiceClientWrapper
At the initial creation of the framework it was intented for bundles to provide a callback which would be called with undefined or the client if the client changed. On top of this `ServiceClientWrapper` was built which allowed bundles to have a separate callback for when a client is available or unavailable. `ServiceClientWrapper` used the previously used callback which in retrospect was just an ugly addition.
This PR refactors it to be more integrated with the framework by completely removing the old callback with it and renaming it to `ServiceProvider` which imo is a way better name.
Also now the `EventEmitter` is now in a variable so that it doesn't feel so bloated with all the `EventEmitter` functions when using InteliSense on it.