Add install.json from nodecg-io-cli to .gitignore
This file is created by the nodecg-io cli when you create a install and saves what you have selected there. In case you create a development install git will say that this file is untracked because it is not checked in and not gitignored.
This file should not be commited so this PR adds it to the `.gitignore` file.