Remove main entry from twitch-auth package.json
`nodecg-io-twitch-auth` has its main file directly in `index.{ts,js}`, which is the default. But it was set to `extension/index` which does not exist and is incorrect. This happend because we I've copied `package.json` from some other package and didn't notice this.
Previous node.js versions simply ignored this but v16 and newer report this as a [runtime error](https://nodejs.org/api/deprecations.html#DEP0128).