Create separate tsconfig for core tests with noEmit enabled
We need to have the test files included in a `tsconfig.json` file for VS Code integration but
we don't want compilation results because jest would pick up
on them (especially `.js` and `.d.ts` files) and try to run them.

To fix this, this commit creates a serparate tsconfig for the tests with noEmit enabled.
This means that VS Code can use this file for TypeScript integration
and we won't emit any compiled files which would mess with our jest setup.