Enforce type checking indexed accesses
Without explicit config TypeScript allows to access a element of a object or array by just accessing it using a index.
But in reality you might get back `undefined` because the object doesn't contain anything with that key or the array isn't long enough.
To force these checks I've typed `ObjectMap` in a way that all values might always be undefined.
This is not good because you might use e.g. `Object.entries` to iterate
a object when using it as a map that ensures you'll only get entries with set values.

TypeScript has a built in solution to this: `noUncheckedIndexedAccess`.
When this compiler setting is enabled it will ensure that you'll check values you get from index access for `undefined`.
This also allows us to use `Object.entries` to get rid of a lot of redudant undefined checks which makes the code a bit cleaner.
Also forces services to do the same because it is nicer to get a error that you might get `undefined` instead of the user catching this
because we never used it in a way that produces `undefined` and the typings didn't contain it.

Also enabling this has brought up some issues with the dashboard code that could have resulted in using `undefined` as a object key.