Implement `localhost` default address for IntelliJ service config
The IntelliJ schema mentions that it defaults to `127.0.0.1:19524` if no address is given.
This makes sense as most will connect the service to the IntelliJ instance running on the same host.
However the schema requires a value and also the code doesn't check for `undefined` and thus doesn't provide the default if the value is missing.
This PR implements the default value by using it if no value is provided and removing `address` from the required fields list of the JSON schema.