Remove JSON schemas from configuration-less services
The curseforge and debug service don't require any configuration and therefore the current JSON schema is empty.
This is unneeded because the schema isn't used to validate the config as there is no config.
To make these services easier we can safely remove the schemas as they are optional anyway.