Don't run docs update action for events caused by dependabot (#296)
All GitHub Actions runs triggered by dependabot have no access to secrets for security reasons.
Because of this our documentation update action is failing when dependabot merges a pull request.
This pr adds a check to the action. Now it will only run if it was not triggered by dependabot.