Move nodecg-types to normal dependencies for debug and nanoleaf service
The debug service exposes the `NodeCG` type with its constructor and nanoleaf with the `NanoleafUtils.retrieveAuthKey` method.
Because of this the generated `.d.ts` files import `nodecg-types` and we need to make it a production dependency as it is needed to be able to use the typings when e.g. adding it as a dependency to a bundle once it has been released.