Downgrade monaco-editor from 0.31.1 to 0.30.1
Reverts d1da66337a6c0c141d174bd5aaa8a4be4baf545a / #377

There is some change with `0.30.0` that results in a build output by webpack that
is incompatible with the Chromium version that is used by custom OBS Docks.

I don't know what the exact problem is yet with this specific update,
this PR is just here to fix this bug till I figure out if we can make
the new version compatible with OBS Docks or not.