Just log a warning when trying automatic login, but nodecg-io is already loaded
Fixes #365.
If automatic login fails and nodecg-io is not loaded afterwards this will still display a error. This happens when e.g. the password set in the automatic login config is wrong.
However if automatic login fails and nodecg-io is loaded afterwards just a warning will be logged with this PR. This is the case when the dashboard reconnects after a restart and loggs back in.