ci: Use timeout option from child_process.spawn
The implications of https://nodejs.org/api/child_process.html#subprocesskillsignal
and the possibility of killing other process, implied using the integrated timeout option.

Also, `stdout` is now collected in a Buffer and is only converted after NodeCG stopped. This is needed in case the data of a single line is split up into to calls which would brake the previous implementation.