Remove curseforge service
The curseforge api behind the curseforge service was deprecated: https://twitter.com/curseforge/status/1522266862322692097?lang=en
There is a new api but from my understanding it is pretty different and
making the client work with it would be a non trivial changed.
@MelanX (creator of the service) said I should remove the service once
the api was shutdown, which was already before a few months.