Remove @types/gapi dependency from googleapis service
The `nodecg-io-googleapis` service depends on the `googleapis` package
for the service implementation.
The service was added in ed7eb2f71807378dae9a5462f7a481df126b2987
with an additional dependency to `@types/gapi` for types.
However the used `googleapis` package provides its own typings out of
the box and `@types/gapi` is for the outdated `gapi` package which is
not used here.

Removes the mentioned unused typing package as it is not used.