Add built assets to tarball of the dashboard
Previously no `files` key was specified in the package.json of the
dashboard.
Because of this only the source files were added to the dashboard v0.3.0
tarball.
This PR updates the package.json to include the built version that is
used by the production installations that are made using the
nodecg-io-cli.

After this PR has been merged it will be backported to the `release/0.3`
branch and v0.3.1 of the dashboard will be released.