Switch from pretty-quick to lint-staged for auto formatting on commit
With prettier v3 pretty-quick does not work anymore.
It hasn't had a new commit for two years so it is unlikely it will
ever support prettier v3.
So this PR migrates the pre-commit hook to use lint-staged
which also supports running eslint on the changed files
before a commit and is maintained.